+++ core/persistence-api/src/main/java/org/apache/syncope/core/persistence/api/dao/AnySearchDAO.java
@@ -37,7 +37,7 @@ public interface AnySearchDAO {
      * @param kind any object
      * @return size of search result
      */
-    int count(
+    long count(
             Realm base,
             boolean recursive,
             Set<String> adminRealms,
